#23 login flow fixed
