+++ src/views/UserLogin.vue
@@ -17,9 +17,6 @@
             :class="{ 'p-invalid': v$.email.$error || check.email.unregistered }"
           />
           <small v-if="v$.email.$error" id="email-help" class="p-error">{{ v$.email.$errors[0].$message }}</small>
-          <small v-else-if="check.email.unregistered" id="email-unregistered" class="p-error">
-            Такого користувача немає в системі
-          </small>
         </div>
         <div class="form-login__item">
           <label for="password">Пароль</label>
@@ -29,7 +26,7 @@
             placeholder="Пароль"
             v-model="password"
             @blur="passwordBlur"
-            :class="{ 'p-invalid': v$.password.$error }"
+            :class="{ 'p-invalid': v$.password.$error || check.password.correct }"
           />
           <small v-if="v$.password.$error" id="password-help" class="p-error">{{
             v$.password.$errors[0].$message
@@ -61,6 +58,7 @@ import {
   emailLastCharsValidator,
   passwordMinLenght,
   passwordMaxLenght,
+  passwordValidator,
 } from '@/utils/validators';
 
 export default defineComponent({
@@ -109,6 +107,7 @@ export default defineComponent({
         requiredValidator,
         passwordMaxLenght,
         passwordMinLenght,
+        passwordValidator,
       },
     };
   },
